# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "ConfigOut": {
        "description": "The limits a client has to know and cannot derive.\n\nEvery value is a setting or a constant a route already enforces, so nothing\nhere is a second statement of a limit: a client that ignores this document\nlearns the same numbers from a `413`, a `409` or a `400 bad_bucket`.",
        "properties": {
          "ack_max": {
            "title": "Ack Max",
            "type": "integer"
          },
          "attachment_buckets": {
            "items": {
              "type": "integer"
            },
            "title": "Attachment Buckets",
            "type": "array"
          },
          "attachment_ttl_days": {
            "title": "Attachment Ttl Days",
            "type": "integer"
          },
          "claim_max": {
            "title": "Claim Max",
            "type": "integer"
          },
          "drain_page_max": {
            "title": "Drain Page Max",
            "type": "integer"
          },
          "envelope_buckets": {
            "items": {
              "type": "integer"
            },
            "title": "Envelope Buckets",
            "type": "array"
          },
          "envelope_ttl_days": {
            "title": "Envelope Ttl Days",
            "type": "integer"
          },
          "mailbox_max_bytes": {
            "title": "Mailbox Max Bytes",
            "type": "integer"
          },
          "max_devicelog_records": {
            "title": "Max Devicelog Records",
            "type": "integer"
          },
          "max_devices_per_user": {
            "title": "Max Devices Per User",
            "type": "integer"
          },
          "send_batch_max": {
            "title": "Send Batch Max",
            "type": "integer"
          },
          "session_token_days": {
            "title": "Session Token Days",
            "type": "integer"
          },
          "signal_buckets": {
            "items": {
              "type": "integer"
            },
            "title": "Signal Buckets",
            "type": "array"
          },
          "voice_configured": {
            "title": "Voice Configured",
            "type": "boolean"
          }
        },
        "required": [
          "envelope_ttl_days",
          "attachment_ttl_days",
          "mailbox_max_bytes",
          "max_devices_per_user",
          "max_devicelog_records",
          "session_token_days",
          "send_batch_max",
          "ack_max",
          "drain_page_max",
          "claim_max",
          "envelope_buckets",
          "attachment_buckets",
          "signal_buckets",
          "voice_configured"
        ],
        "title": "ConfigOut",
        "type": "object"
      },
...
      "ErrorOut": {
        "description": "The error envelope. `detail` is a string on every code but\n`invalid_request`, where it maps a field path to the messages that failed.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "additionalProperties": {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                "type": "object"
              }
            ],
            "title": "Detail"
          }
        },
        "required": [
          "code",
          "detail"
        ],
        "title": "ErrorOut",
        "type": "object"
      },
...
      "RelayCredentialOut": {
        "description": "One ephemeral coturn credential.\n\n`urls` is the configured `turn:` list, in the order the operator wrote it.\n`username` is the Unix expiry timestamp, a colon and sixteen random bytes as\nURL-safe base64. `credential` is standard base64 of HMAC-SHA1 over that\nusername. `expires_in` is the seconds the pair stays good for, counted from\nthe moment it was minted.",
        "properties": {
          "credential": {
            "title": "Credential",
            "type": "string"
          },
          "expires_in": {
            "title": "Expires In",
            "type": "integer"
          },
          "urls": {
            "items": {
              "type": "string"
            },
            "title": "Urls",
            "type": "array"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "urls",
          "username",
          "credential",
          "expires_in"
        ],
        "title": "RelayCredentialOut",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "communication platform",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/config": {
      "get": {
        "description": "The limits the client contract names, read from the settings and the\nconstants the routes enforce rather than restated here.\n\nIt exists because the client cannot otherwise state them. `ENVELOPE_TTL_DAYS`\nis an operator setting, and a client that tells a user how long an undelivered\nmessage survives has to be told what it is. Authenticated, because none of it\nis a secret and none of it is anyone's business before they hold a session.\n\nIt touches no row, so it needs no unit of work.",
        "operationId": "config",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "`unauthenticated`, `invalid_token`, `token_revoked`"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "`scope_forbidden`"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "`throttled`"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "`server_error`"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "`unavailable`"
          }
        },
        "summary": "Config",
        "tags": [
          "core"
        ]
      }
    },
...
    "/api/v1/me/relay": {
      "post": {
        "description": "Mint an ephemeral coturn credential for the calling device.\n\nSafe to repeat: every call mints a fresh username, and a credential already\nissued stays good until its own expiry, so a client that retries a timed-out\nrequest holds two working credentials rather than none.\n\nAn unconfigured deployment answers `503 voice_unconfigured` rather than a\ncredential no relay would accept. It is not a backoff: `TURN_URLS` is empty,\nso this server does not do voice, and a retry will not make a relay appear.\nThe route reads the setting and never reaches coturn, so a relay that is\nconfigured but down is a `200` and a call that fails to connect.",
        "operationId": "mint_relay_credential",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RelayCredentialOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "`unauthenticated`, `invalid_token`, `token_revoked`"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "`scope_forbidden`"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "`throttled`"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "`server_error`"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "`voice_unconfigured`, `unavailable`"
          }
        },
        "summary": "Mint Relay Credential",
        "tags": [
          "realtime"
        ]
      }
    },
...
  }
}